+++ CHANGES.md
@@ -1,3 +1,8 @@
+Version 2.9.0.9
+
+ - Support: RENAME and RENAMENX
+ - SADD will raise an exception if an empty list is passed to it
+
 Version 2.9.0.8
 
  - Add inclusive syntax (parenthesis) support for zero sets ZRANGEBYSCORE, ZREVRANGEBYSCORE  & ZREMRANGEBYSCORE